forge fmt and remove unnecesary comments

+++ src/SynthetixSafeModule.sol
@@ -10,13 +10,13 @@ import "./SignatureDecoder.sol";
 import "forge-std/console.sol";
 
 contract SynthetixSafeModule is IGuard, SignatureDecoder {
-
     // required constants from gnosis
 
     // keccak256(
     //     "EIP712Domain(uint256 chainId,address verifyingContract)"
     // );
-    bytes32 private constant DOMAIN_SEPARATOR_TYPEHASH = 0x47e79534a245952e8b16893a336b85a3d9ea9fa8c573f3d803afb92a79469218;
+    bytes32 private constant DOMAIN_SEPARATOR_TYPEHASH =
+        0x47e79534a245952e8b16893a336b85a3d9ea9fa8c573f3d803afb92a79469218;
 
     // keccak256(
     //     "SafeTx(address to,uint256 value,bytes data,uint8 operation,uint256 safeTxGas,uint256 baseGas,uint256 gasPrice,address gasToken,address refundReceiver,uint256 nonce)"
@@ -24,21 +24,21 @@ contract SynthetixSafeModule is IGuard, SignatureDecoder {
     bytes32 private constant SAFE_TX_TYPEHASH = 0xbb8310d486368db6bd6f849402fdd73ad53d316b5a4b2644ad6efe0f941286d8;
 
     error SafeCallFailed(address safe, bytes attemptedCall);
-    error InsufficientSigners(bytes32 group, uint required, uint provided);
+    error InsufficientSigners(bytes32 group, uint256 required, uint256 provided);
 
     IElectionModule public electionSystem;
 
     ISafe public pdaoSafe;
 
     // the number of votes required from the pdao to pass
-    uint public pdaoThreshold;
+    uint256 public pdaoThreshold;
 
     constructor(IElectionModule _electionSystem, ISafe _pdaoSafe) {
         electionSystem = _electionSystem;
         pdaoSafe = _pdaoSafe;
     }
 
-    function setPdaoThreshold(uint threshold) external {
+    function setPdaoThreshold(uint256 threshold) external {
         require(msg.sender == address(pdaoSafe), "pdao only");
 
         pdaoThreshold = threshold;
@@ -54,46 +54,49 @@ contract SynthetixSafeModule is IGuard, SignatureDecoder {
         address[] memory electedCouncilSigners = electionSystem.getCouncilMembers();
         address[] memory pdaoSigners = pdaoSafe.getOwners();
 
-        uint requiredSigners = electedCouncilSigners.length / 2 + 1 + pdaoThreshold;
+        uint256 requiredSigners = electedCouncilSigners.length / 2 + 1 + pdaoThreshold;
 
         // remove all signers currently on the target safe except one (because gnosis does not allow no signers)
         address[] memory oldSigners = targetSafe.getOwners();
-        for (uint i = 1;i < oldSigners.length;i++) {
-            execOnSafe(
-                targetSafe, 
-                abi.encodeWithSelector(ISafe.removeOwner.selector, oldSigners[0], oldSigners[i], 1)
-            );
+        for (uint256 i = 1; i < oldSigners.length; i++) {
+            execOnSafe(targetSafe, abi.encodeWithSelector(ISafe.removeOwner.selector, oldSigners[0], oldSigners[i], 1));
         }
 
         // add new signers to the target safe
-        uint addedSigners = 0;
-        for (uint i = 0; i < pdaoSigners.length; i++) {
+        uint256 addedSigners = 0;
+        for (uint256 i = 0; i < pdaoSigners.length; i++) {
             if (oldSigners.length > 0 && pdaoSigners[i] == oldSigners[0]) {
                 oldSigners = new address[](0);
-            }
-            else {
+            } else {
                 execOnSafe(
-                    targetSafe, 
-                    abi.encodeWithSelector(ISafe.addOwnerWithThreshold.selector, pdaoSigners[i], requiredSigners < ++addedSigners ? requiredSigners : addedSigners)
+                    targetSafe,
+                    abi.encodeWithSelector(
+                        ISafe.addOwnerWithThreshold.selector,
+                        pdaoSigners[i],
+                        requiredSigners < ++addedSigners ? requiredSigners : addedSigners
+                    )
                 );
             }
         }
 
-        for (uint i = 0; i < electedCouncilSigners.length; i++) {
+        for (uint256 i = 0; i < electedCouncilSigners.length; i++) {
             if (oldSigners.length > 0 && electedCouncilSigners[i] == oldSigners[0]) {
                 oldSigners = new address[](0);
-            }
-            else if (!targetSafe.isOwner(electedCouncilSigners[i])) {
+            } else if (!targetSafe.isOwner(electedCouncilSigners[i])) {
                 execOnSafe(
-                    targetSafe, 
-                    abi.encodeWithSelector(ISafe.addOwnerWithThreshold.selector, electedCouncilSigners[i], requiredSigners < ++addedSigners ? requiredSigners : addedSigners)
+                    targetSafe,
+                    abi.encodeWithSelector(
+                        ISafe.addOwnerWithThreshold.selector,
+                        electedCouncilSigners[i],
+                        requiredSigners < ++addedSigners ? requiredSigners : addedSigners
+                    )
                 );
             }
         }
 
         if (oldSigners.length > 0) {
             execOnSafe(
-                targetSafe, 
+                targetSafe,
                 abi.encodeWithSelector(ISafe.removeOwner.selector, pdaoSigners[0], oldSigners[0], requiredSigners)
             );
         }
@@ -101,7 +104,7 @@ contract SynthetixSafeModule is IGuard, SignatureDecoder {
 
     /**
      * Ensures that a minimum number of pdao signers have signed this txn
-     */ 
+     */
     function checkTransaction(
         address to,
         uint256 value,
@@ -115,7 +118,6 @@ contract SynthetixSafeModule is IGuard, SignatureDecoder {
         bytes memory signatures,
         address msgSender
     ) external {
-
         bytes memory txHashData;
         {
             txHashData = ISafe(msg.sender).encodeTransactionData(
@@ -134,19 +136,19 @@ contract SynthetixSafeModule is IGuard, SignatureDecoder {
                 ISafe(msg.sender).nonce() - 1
             );
         }
-        
+
         address[] memory electedCouncilSigners = electionSystem.getCouncilMembers();
         address[] memory pdaoSigners = pdaoSafe.getOwners();
 
-        uint electedCount = 0;
-        uint pdaoCount = 0;
+        uint256 electedCount = 0;
+        uint256 pdaoCount = 0;
 
         uint8 v;
         bytes32 r;
         bytes32 s;
         address curOwner;
 
-        for (uint j = 0;j < signatures.length / 65;j++) {
+        for (uint256 j = 0; j < signatures.length / 65; j++) {
             (v, r, s) = signatureSplit(signatures, j);
             if (v == 0) {
                 revert("Contract signatures are not supported by this module");
@@ -155,7 +157,9 @@ contract SynthetixSafeModule is IGuard, SignatureDecoder {
                 curOwner = msgSender;
             } else if (v > 30) {
                 // To support eth_sign and similar we adjust v and hash the transferHashData with the Ethereum message prefix before applying ecrecover
-                curOwner = ecrecover(keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", keccak256(txHashData))), v - 4, r, s);
+                curOwner = ecrecover(
+                    keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", keccak256(txHashData))), v - 4, r, s
+                );
             } else {
                 // Use ecrecover with the messageHash for EOA signatures
                 curOwner = ecrecover(keccak256(txHashData), v, r, s);
@@ -164,7 +168,7 @@ contract SynthetixSafeModule is IGuard, SignatureDecoder {
             require(curOwner != address(0), "curOwner != address(0)");
 
             if (electedCount < electedCouncilSigners.length / 2 + 1) {
-                for (uint i = 0;i < electedCouncilSigners.length;i++) {
+                for (uint256 i = 0; i < electedCouncilSigners.length; i++) {
                     if (electedCouncilSigners[i] == curOwner) {
                         electedCount++;
                         break;
@@ -173,7 +177,7 @@ contract SynthetixSafeModule is IGuard, SignatureDecoder {
             }
 
             if (pdaoCount < pdaoThreshold) {
-                for (uint i = 0;i < pdaoSigners.length;i++) {
+                for (uint256 i = 0; i < pdaoSigners.length; i++) {
                     if (pdaoSigners[i] == curOwner) {
                         pdaoCount++;
                         break;
@@ -194,12 +198,10 @@ contract SynthetixSafeModule is IGuard, SignatureDecoder {
     function checkAfterExecution(bytes32 txHash, bool success) external {}
 
     function supportsInterface(bytes4 interfaceId) external view virtual override returns (bool) {
-        return
-            interfaceId == type(IGuard).interfaceId || // 0xe6d7a83a
-            interfaceId == type(IERC165).interfaceId; // 0x01ffc9a7
+        return interfaceId == type(IGuard).interfaceId // 0xe6d7a83a
+            || interfaceId == type(IERC165).interfaceId; // 0x01ffc9a7
     }
 
-    
     // functions ripped from gnosis safe contract needed for verification:
 
     /// @dev Returns the chain id used by this contract.
@@ -213,12 +215,7 @@ contract SynthetixSafeModule is IGuard, SignatureDecoder {
     }
 
     function execOnSafe(ISafe safe, bytes memory call) internal {
-        bool success = safe.execTransactionFromModule(
-            address(safe),
-            0,
-            call,
-            Enum.Operation.Call
-        );
+        bool success = safe.execTransactionFromModule(address(safe), 0, call, Enum.Operation.Call);
 
         if (!success) {
             revert SafeCallFailed(address(safe), call);
